+++ examples/report-generation-example/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ballerine/report-generation-example
 
+## 0.2.18
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/react-pdf-toolkit@1.2.35
+
+## 0.2.17
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/react-pdf-toolkit@1.2.34
+
+## 0.2.16
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/react-pdf-toolkit@1.2.33
+
 ## 0.2.15
 
 ### Patch Changes